requirements.txt
asgiref==3.8.1
Django==5.1.3
dnspython==2.7.0
paho-mqtt==2.1.0
pymongo==4.10.1
sqlparse==0.5.1
